7-fileanalysis/requirements.txt
Flask==0.12
kafka-python
